# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "STD Plus API",
    "version": "1.0.0",
    "description": "Backend API for STD Plus — materials, inventory, orders, and workflow management."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "message": { "type": "string" },
          "data": {}
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "message": { "type": "string" }
        }
      },
      "ValidationErrorResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "message": { "type": "string", "example": "Validation failed" },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": { "type": "string" },
                "message": { "type": "string" }
              }
            }
          }
        }
      },
      "DeleteManyBody": {
        "type": "object",
        "required": ["ids"],
        "properties": {
          "ids": {
            "type": "array",
            "items": { "type": "string" },
            "minItems": 1
          }
        }
      },
      "DeleteManyResult": {
        "type": "object",
        "properties": {
          "deletedCount": { "type": "integer" }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": { "type": "integer", "example": 1 },
          "limit": { "type": "integer", "example": 20 },
          "total": { "type": "integer", "example": 45 },
          "totalPages": { "type": "integer", "example": 3 }
        }
      },
...
      "StationTemplate": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "name": { "type": "string" },
          "uiSchema": { "type": "object" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "Station": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "name": { "type": "string" },
          "templateId": { "oneOf": [{ "type": "string" }, { "$ref": "#/components/schemas/StationTemplate" }] },
          "colorId": { "type": "string", "enum": ["sky", "blue", "violet", "pink", "red", "orange", "yellow", "green", "teal", "slate"], "default": "sky", "description": "Station color identifier for UI display" },
          "status": { "type": "string", "enum": ["online", "offline", "maintenance"], "default": "offline" },
          "isLaminateStation": { "type": "boolean", "default": false, "description": "If true, this station is where the `laminate` scan runs: active sheets merge and one **survivor** sheet keeps its QR; others and the dormant parent row become `merged_into`" },
          "notes": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "StationCreateBody": {
        "type": "object",
        "required": ["name", "templateId"],
        "properties": {
          "name": { "type": "string", "minLength": 1 },
          "templateId": { "type": "string", "minLength": 1 },
          "colorId": { "type": "string", "enum": ["sky", "blue", "violet", "pink", "red", "orange", "yellow", "green", "teal", "slate"] },
          "status": { "type": "string", "enum": ["online", "offline", "maintenance"] },
          "isLaminateStation": { "type": "boolean", "description": "Mark as lamination station (`laminate` action; one sheet survives as the product QR)" },
          "notes": { "type": "string" }
        }
      },
...
    },
    "parameters": {
...
      "PageParam": {
        "name": "page",
        "in": "query",
        "schema": { "type": "integer", "minimum": 1, "default": 1 },
        "description": "Page number (1-indexed)"
      },
      "LimitParam": {
        "name": "limit",
        "in": "query",
        "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 },
        "description": "Items per page (clamped to 1–100)"
      },
      "SortParam": {
        "name": "sort",
        "in": "query",
        "schema": { "type": "string", "default": "-createdAt" },
        "description": "Sort field. Prefix with - for descending (e.g. -createdAt, name)"
      }
    },
    "responses": {
      "401": {
        "description": "Not authenticated",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
      "400": {
        "description": "Validation failed",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ValidationErrorResponse" }
          }
        }
      },
      "403": {
        "description": "Not authorized for this action",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "409": {
        "description": "Cannot delete — referenced by other records",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/stations": {
      "get": {
        "tags": ["Stations"],
        "summary": "List all stations",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/PageParam" },
          { "$ref": "#/components/parameters/LimitParam" },
          { "$ref": "#/components/parameters/SortParam" }
        ],
        "responses": {
          "200": { "description": "List of stations", "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/SuccessResponse" }, { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Station" } }, "pagination": { "$ref": "#/components/schemas/Pagination" } } }] } } } },
          "401": { "$ref": "#/components/responses/401" }
        }
      },
      "post": {
        "tags": ["Stations"],
        "summary": "Create a station",
        "description": "Requires `stations:manage` permission.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/StationCreateBody" } } } },
        "responses": {
          "201": { "description": "Station created", "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/SuccessResponse" }, { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/Station" } } }] } } } },
          "400": { "$ref": "#/components/responses/400" },
          "401": { "$ref": "#/components/responses/401" },
          "403": { "$ref": "#/components/responses/403" }
        }
      },
      "delete": {
        "tags": ["Stations"],
        "summary": "Delete multiple stations",
        "description": "Requires `stations:manage` permission.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DeleteManyBody" } } } },
        "responses": {
          "200": { "description": "Stations deleted", "content": { "application/json": { "schema": { "allOf": [{ "$ref": "#/components/schemas/SuccessResponse" }, { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/DeleteManyResult" } } }] } } } },
          "401": { "$ref": "#/components/responses/401" },
          "403": { "$ref": "#/components/responses/403" },
          "409": { "$ref": "#/components/responses/409" }
        }
      }
    },
...
  },
...
}